+++ src/hooks.js
@@ -21,6 +21,12 @@ export const HOOKS = {
     createRxQuery: [],
     createRxDocument: [],
     /**
+     * runs after a RxDocument is created,
+     * async
+     * @type {Array}
+     */
+    postCreateRxDocument: [],
+    /**
      * runs before a pouchdb-instance is created
      * gets pouchParameters as attribute so you can manipulate them
      * {