+++ src/rx-document.js
@@ -452,6 +452,12 @@ export class RxDocument {
     }
 }
 
+/**
+ * createas an RxDocument from the jsonData
+ * @param  {RxCollection} collection
+ * @param  {[type]} jsonData   [description]
+ * @return {RxDocument}
+ */
 export function create(collection, jsonData) {
     if (
         jsonData[collection.schema.primaryPath] &&