+++ test/unit/plugin.test.js
@@ -3,6 +3,8 @@
  */
 
 import assert from 'assert';
+import AsyncTestUtil from 'async-test-util';
+
 import config from './config';
 import RxDB from '../../dist/lib/index';
 import * as util from '../../dist/lib/util';
@@ -132,6 +134,22 @@ describe('plugin.test.js', () => {
             const col = await humansCollection.create(5);
             const doc = await col.findOne().exec();
             assert.equal(doc.foo, 'bar_createRxDocument');
+            col.database.destroy();
+        });
+        it('postCreateRxDocument', async () => {
+            const plugin = {
+                rxdb: true,
+                hooks: {
+                    postCreateRxDocument: async (doc) => {
+                        await AsyncTestUtil.wait(100);
+                        doc.postCreateFoo = 'postCreateRxDocument worked!';
+                    }
+                }
+            };
+            RxDB.plugin(plugin);
+            const col = await humansCollection.create(5);
+            const doc = await col.findOne().exec();
+            assert.equal(doc.postCreateFoo, 'postCreateRxDocument worked!');
             col.database.destroy();
         });
         it('preCreatePouchDb', async () => {