+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Features:
   - Added [options-parameter](https://pubkey.github.io/rxdb/plugins.html#options)
+  - Added `postCreateRxDocument` [plugin-hook](https://github.com/pubkey/rxdb/blob/master/src/hooks.js)
 
 Bugfixes:
   - Query-Cache not used when declaring queries without mango-chain